+++ source/Axiom/Axiom/Codecs/Video/AV1.cs
@@ -218,7 +218,7 @@ namespace Axiom
             vm.FPS_IsEnabled = true;
 
             // Scaling ComboBox
-            vm.Scaling_IsEnabled = true;
+            vm.ScalingAlgorithm_IsEnabled = true;
 
             // Crop
             vm.Crop_IsEnabled = true;