+++ source/Axiom/Axiom/Codecs/Video/x264.cs
@@ -270,7 +270,7 @@ namespace Axiom
             vm.Video_Optimize_IsEnabled = true;
 
             // Scaling ComboBox
-            vm.Scaling_IsEnabled = true;
+            vm.ScalingAlgorithm_IsEnabled = true;
 
             // Crop
             vm.Crop_IsEnabled = true;